move fetching outside the component
